perf: split borrow action hooks for DeFi dialog
